Back-End/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
alembic
python-jose[cryptography]
pydantic
pydantic-settings
passlib[bcrypt]
python-dotenv
email-validator